requirements.txt
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn>=0.30.0
torch>=2.4.0
torchvision>=0.19.0
pillow>=10.0.0
huggingface_hub>=0.25.0
faiss-cpu>=1.8.0
pydantic>=2.0.0
numpy>=1.26.0